Implementación HU6 2
En el commit anterior hubo una equivocación en el nombre, en verdad era la implementación del HU5, no del HU6. En cambio, ahora sí se sube la implementación del HU6, la cual corresponde a calcular la plantilla de pagos de arancel.

+++ src/main/java/com/example/evaluacion1/entities/CuotaEntity.java
@@ -24,8 +24,8 @@ public class CuotaEntity {
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date fechaVencimiento;
     private boolean pagado;
-    private float descuento;
-    private float interes;
+    private double descuento;
+    private double interes;
     private int montoCuota;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date fechaPago;
